requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-filter==25.1
drf-yasg==1.21.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Pillow==11.1.0
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==5.0.0
factory-boy==3.3.0
django-debug-toolbar==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.4.0
redis==5.0.2
django-cors-headers==4.3.1
dj-database-url==2.1.0 